+++ lambda/interrupts.h
@@ -24,7 +24,7 @@ uint32_t getInts(void);
  * Adds the given number of interrupts to the interrupt counter
  * used as timebase. Useful only for simulation.
  */
-void addInts(uint8_t add);
+void addInts(uint32_t add);
 
 /**
  * Returns the time in about seconds since last reset.
